// paaq-client/src/app/api/v1/model/sensor.ts
/**
 * Aquaponic Sensors DB API
 * An API to manage the sensors in the Aquaponic system
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: akcaden1@students.zhaw.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Sensor { 
    id?: number;
    name: string;
    serialNrDevice: string;
    deviceType: string;
    unit: string;
    tank: string;
    system: string;
    startingPos: number;
    length: number;
    dataType: string;
    modBusPort: number;
    serial_port: string;
    alarmUpperLimit: number;
    alarmLowerLimit: number;
    alaramActive?: number;
}